Компилируется Keil Device

+++ sources/Device/src/Utils/Values.cpp
@@ -452,7 +452,7 @@ static char *FloatToString(float value, bool alwaysSign, int numDigits, char buf
     }
     
     bool signExist = alwaysSign || value < 0; //-V2564
-    while (std::strlen(bufferOut) < static_cast<size_t>(numDigits + (signExist ? 2 : 1))) //-V2513
+    while (std::strlen(bufferOut) < static_cast<uint>(numDigits + (signExist ? 2 : 1))) //-V2513
     {
         std::strcat(bufferOut, "0"); //-V2513
     }
